get best approximation

+++ utils/math/math.c
@@ -2,15 +2,15 @@
 #include <stdio.h>
 #include <math.h>
 
-float calculateIntegrate(float f(float), float startPos, float endPos)
+double calculateIntegrate(double f(double), double startPos, double endPos)
 {
 #define N 100000
 
-    float totalArea = 0;
-    float width;
+    double totalArea = 0;
+    double width;
 
     width = (endPos - startPos) / N;
-    for (float i = 1; i <= N; i++)
+    for (double i = 1; i <= N; i++)
     {
         totalArea = totalArea + f(startPos + width * i) * width;
     }
